requirements.txt
scikit-learn~=1.5.0
scipy~=1.13.1
tqdm~=4.66.4
torch~=1.13.0+cu116
numpy~=1.20.3
torchvision~=0.18.0+cu121
opencv-python~=4.10.0.82
SimpleITK~=2.3.1
natsort~=8.4.0
matplotlib~=3.8.4
timm~=0.3.2
MedPy
monai~=1.3.0
Pillow~=10.2.0
nibabel~=5.2.1